[doc] Modify the `allow_no_indices` docstring
The `allow_no_indices` parameter in the methods `get`, `delete`, and
`exists`, of the `indices.IndicesClient` was misleading by reporting an
incorrect default value (report false, when it was actually true).

This commit modifies the docstring of this parameter in the three
aforementioned methods, by replacing it with the docstring of this
parameter from the other methods of the codebase. The new one does not
report a default value, because the elasticsearch-py does not actually
set a default in case of absence when performing the HTTP request.