Backport aiohttp conditional HEAD bug workaround
In older versions of aiohttp, HEAD requests don't mark the connection as
not reusable, which is why GET was used instead. But this means
`indices.exists` becomes `indices.get` which can be prohibitively slow
on indices with large mappings, so we want to use `HEAD` for aiohttp
versions that don't have this bug.

This backports the following pull requests to the 7.17 branch to make it
easier for Rally to upgrade to elasticsearch-py 8.x:

 * https://github.com/elastic/elastic-transport-python/pull/55
 * https://github.com/elastic/elastic-transport-python/pull/58