Replace unsafe API method for temporary file creation
In file: generate-examples.py, there is a method that creates a temporary file using an unsafe API mktemp. The use of this is discouraged in the Python documentation (https://docs.python.org/3/library/tempfile.html#tempfile.mktemp). We suggested that a temporary file should be created using NamedTemporaryFile(https://docs.python.org/3/library/tempfile.html#tempfile.NamedTemporaryFile) which is a safe API.