Add back inference.inference API (#2873)
* Add back inference.inference API

This function was removed from the specification in favor of one API per
provider and task type, but the existing function was stable and widely used in
Python. Still, we mark it as deprecated to encourage users to migrate to the
new APIs.

* Fix lint

(cherry picked from commit eac539dcc301b6062716f1540dc0d1efe888671c)